Refactor : 파티 목록 관련 이름 변경
useChat 훅과 일관성을 맞추기 위해 파일 이름을 변경하였습니다.
useParty => usePartyDetail
useParties => useParty
usePartiesUpdate => usePartyUpdate
