+++ src/hooks/useParty.jsx
@@ -1,23 +1,34 @@
 import { useEffect, useState } from 'react';
-import { SERVER_URL } from '../config';
+import { PARTY_PAGE_SIZE, SERVER_URL } from '../config';
+import { convertOptionToParam } from '../utils/converter';
 
-function useParty(id) {
-  const [party, setParty] = useState();
+function useParty(search) {
   const [parties, setParties] = useState([]);
+  const [option, setOption] = useState({
+    search: search || '',
+    OPEN: true,
+    WISHLIST: false,
+    place: {
+      SINCHON: true,
+      YEONHUI: true,
+      CHANGCHEON: true,
+    },
+  });
 
   useEffect(() => {
     const fetchParty = async () => {
-      const res = await fetch(`${SERVER_URL}/parties/${id}`);
+      const params = convertOptionToParam(option);
+      const URL = `${SERVER_URL}/parties?${params.join('&')}&pageSize=${PARTY_PAGE_SIZE}`;
+      const res = await fetch(URL);
       const json = await res.json();
 
-      setParties(json.data.parties);
-      setParty({ ...json.data, id });
+      setParties([...json.data.parties]);
     };
 
     fetchParty();
-  }, [id]);
+  }, [option]);
 
-  return { party, parties };
+  return { parties, setParties, option, setOption };
 }
 
 export default useParty;
